refactor: split prov.model into a package
Split the monolithic src/prov/model.py (3099 lines) into a package with
three implementation modules, re-exporting every public name at its
historic prov.model location. Pure code moves — no behaviour changes.

- records.py: datatype helpers, Literal, exceptions, ProvRecord and every
  concrete element/relation class, and the PROV_REC_CLS registry. References
  ProvBundle only under TYPE_CHECKING (no runtime dependency on bundle.py).
- namespaces.py: NamespaceManager and DEFAULT_NAMESPACES.
- bundle.py: ProvBundle, ProvDocument, sorted_attributes.
- __init__.py: explicit `X as X` re-exports of every public name plus
  `from prov.constants import *`, keeping dir(prov.model) identical to the
  pre-split module. The submodule attributes are removed from the package
  namespace so the frozen public surface is unchanged.

Import graph: records -> (namespaces, bundle); namespaces -> records;
bundle -> (records, namespaces). No runtime cycles.

Update the F403/F405 per-file-ignore path and the CLAUDE.md architecture
section to match the new layout.
