+++ pyproject.toml
@@ -136,7 +136,7 @@ combine-as-imports = true  # keep multi-name `X as X` re-export blocks compact
 [tool.ruff.lint.per-file-ignores]
 # These modules intentionally re-export everything from another module with
 # `import *`; pyflakes can't tell which of the resulting names are "used".
-"src/prov/model.py" = ["F403", "F405"]
+"src/prov/model/__init__.py" = ["F403", "F405"]
 "src/prov/serializers/provjson.py" = ["F403", "F405"]
 "src/prov/serializers/provxml.py" = ["F403", "F405"]
 "src/prov/tests/test_statements.py" = ["F403", "F405"]
